Moved summary definitions to own file
